collectors/instagram_bot/requirements.txt
Scrapy==2.11.0
supabase==2.3.0
schedule==1.2.1
python-dotenv==1.0.0
